Streamline rtc_sentinel.py

+++ hyp3_gamma/util.py
@@ -1,4 +1,3 @@
-import glob
 import logging
 import os
 from zipfile import ZipFile
@@ -9,21 +8,21 @@ from hyp3lib.scene import get_download_url
 log = logging.getLogger(__name__)
 
 
-def find_and_remove(directory, file_pattern):
-    pattern = os.path.join(directory, file_pattern)
-    for filename in glob.glob(pattern):
-        logging.info(f'Removing {filename}')
-        os.remove(filename)
-
-
 def get_granule(granule):
     download_url = get_download_url(granule)
-    zip_file = download_file(download_url)
+    zip_file = download_file(download_url, chunk_size=10485760)
+    safe_dir = unzip_granule(zip_file, remove=True)
+    return safe_dir
+
+
+def unzip_granule(zip_file, remove=False):
     log.info(f'Unzipping {zip_file}')
     with ZipFile(zip_file) as z:
         z.extractall()
-    os.remove(zip_file)
-    return f'{granule}.SAFE'
+        safe_dir = next(item.filename for item in z.infolist() if item.is_dir() and item.filename.endswith('.SAFE/'))
+    if remove:
+        os.remove(zip_file)
+    return safe_dir.strip('/')
 
 
 def earlier_granule_first(g1, g2):
